pyproject.toml
[tool.poetry]
name = "flight-booking"
version = "0.1.0"
description = ""
authors = ["rahoofachulliyil <rahoofachulliyil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = ">=3.2"
djangorestframework = ">=3.12"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
